Add image saving, image data saving to file. Prepare to add such code to 'get_{image_type}' funcs.
This commit largelt comprises of test functions that hold the basis for the future implementation of such mechanisms in functions such as the get_{image_type} functions. It is in perparation for the full data gathering process. This commit also prepares for the solving of the look_at problem, explained in the note at the end of the main_camera.gd file.