Integrate .RDCconfig parsing in tests.py.
This way, we no longer need to change singular values in tests.py itself every time we want to change something, but can rather just change a specific value once in the config of choice. This will soon be integrated in other parts of the project where such info in the config is needed.