First (potentially) working version of data camera version 2.
The camera is actually doing the basic movements and rotating as it should at this point, even though it needs to be verified. It seems (or looks like) that the camera doesn't move far enough to the other side of the cline, as if there aren't enough z steps per row (which is mathematically weird). Probably a bug.