Add progress bar to training & add greater complexity to model
The model still struggles to predict the rotation value when the camera is close to the dock. This is because the camera distorts the dock at the edges, leading to difficult prediction of the rotation value. Because of this, we most likely need an indicator for the reverse line point, which is what we will experiment with next using Godot.