Train on new data set with FOV of 41.2 degrees & adjust model saving and loading in train/inference.
The theory of larger FOV's disrupting model learning capacity/model proficiency at predicting closer images was in fact true. By setting the FOV to a smaller value, the images are more flat, so no more rounding of the dock happens when near the dock. The loss value when trained on smaller FOV value images goes down much more, and the model converges much faster (around 1.75x).

Although the FOV's with this commit are set to 41.2 for the data camera and the robot camera, a larger FOV is most likely appropriate, as right now the image from the camera seems closer than it really is (need to verify this?). We will make the FOV larger most likely, but not as large as the previous 75, though.