Modify data processing, train, and inference scripts to support bimodel functionality
We can actually use the same train.py script to train both models, but as of now the two models both predict a distance cline, which still needs to be verified whether or not the second model having that is necessary. For now, we will keep the scripts under the 'model1' directory, though this may change if it is verified that these scripts being used for both models is the best option.