Add new test that makes sure each image in the data is different & modify training hyperparams
The new data test ensures that each image is different. This is needed because Godot had for some reason gathered the same image for around 630 iterations without changing, probably a Godot bug or process that had stopped during data gathering. The model2 accuracy is super high, with an average difference between the real and pred of .15!