Remove deprecated notes & add target 1 to the scene.
The notes that were there before described the method I was about to
create, but then I realized that it had a flaw and that it would not
work as planned. Thinking it through always helps. Now, however, I have
created a new method that is more realistic than the new one proposed in
the notes removed that should actually work.