Add initial working version of inference script!
It is still a rough implementation, but inference works and best of all,
our simple model trained in 20 minutes on the macbook works, without any
finetuning, first try!! This is a great sign of proper convergence and
good data. However, we still need to fix the issue that the data needs
to be preprocessed/preloaded into tensors, rather than loaded as tensors
during training and/or inferencing.