Add image loading limiter to data function to load specific amounts of images
This way, when we want to do some quick tests or verify certain
procedures during inference, we don't need to wait as long for the
images to load. Rather, we only load the amount necessary. Note: The
default value of the n_images_cls (num images per class) is -1, which
works as a number that never triggers the "break" command of the for
loop getting the data. So, the default is to just load all of the images
available for each class.