Implement action 3 moving forward a fixed amount in virtual robot
The robot works now! However, it is still a bit rough when it targets
the dock. I will have to look into this. It does not move forward
perfectly when starting centered with the dock after model/target
switch, so we might have to implement something like an ultra sonic
sensor to stop a fixed movement forward after the action is 2 (move
forward) to move toward the dock. We will see!