Add custom augmentation to images for training in an attempt to improve training
The model, trained with these parameters and augmentations, is still not
doing entirely well when targeting the dock (red target). I have to look
at this further. It may be due to the lack of images that are up-close,
facing the dock, as we are using images centered on target 1. On a
bigger scale, it may be due to the fact that the model is not complex
enough. We will experiment with this further. The virtual robot does
dock successfully, just not smoothly at the final part.