Add explanation of new method (read commit description)
The new document 'the_new_method.md' is a condensed yet moderately
detailed explanation of the current method used in the project in order
to dock the virtual robot. The new method can of course be used in the
real world, which is what makes it great, unlike the old method. You can
read more about it in the document committed here. A much more detailed
and polished version of such documentation is being created under the
'paper' directory, which I plan to publish to HuggingFace as soon as it
is finished!