[paper] Fix diagram widths to be more compatible with formatting & make improvements to grammar and info
The widths should be fine now. We will see what happens when I add more
diagrams, if the formatting gets messed up or not. In terms of paper
status, it is getting worked on still, and soon it will be published to
arXiv. I need to get endorsed by someone first, so that may take a
moment but it shouldn't be too much of a hassle!